DRY out the code
