+++ front/src/modules/companies/table/components/TableActionBarButtonDeleteCompanies.tsx
@@ -1,18 +1,11 @@
 import { MutableRefObject } from 'react';
 import { getOperationName } from '@apollo/client/utilities';
-import { useRecoilValue } from 'recoil';
 
 import { GET_COMPANIES } from '@/companies/queries';
+import { useDeleteItems } from '@/people/hooks/deleteItems';
 import { GET_PIPELINES } from '@/pipeline/queries';
-import { filtersScopedState } from '@/ui/filter-n-sort/states/filtersScopedState';
-import { Filter } from '@/ui/filter-n-sort/types/Filter';
 import { IconTrash } from '@/ui/icon/index';
-import { useSnackBar } from '@/ui/snack-bar/hooks/useSnackBar';
 import { EntityTableActionBarButton } from '@/ui/table/action-bar/components/EntityTableActionBarButton';
-import { useResetTableRowSelection } from '@/ui/table/hooks/useResetTableRowSelection';
-import { selectedRowIdsSelector } from '@/ui/table/states/selectedRowIdsSelector';
-import { TableContext } from '@/ui/table/states/TableContext';
-import { useRecoilScopedState } from '@/ui/utilities/recoil-scope/hooks/useRecoilScopedState';
 import { useDeleteManyCompaniesMutation } from '~/generated/graphql';
 
 export function TableActionBarButtonDeleteCompanies({
@@ -20,22 +13,6 @@ export function TableActionBarButtonDeleteCompanies({
 }: {
   timerRef: MutableRefObject<NodeJS.Timeout | null>;
 }) {
-  const selectedRowIds = useRecoilValue(selectedRowIdsSelector);
-
-  const { enqueueSnackBar } = useSnackBar();
-  const duration = 5000;
-
-  function handleCloseSnackbar() {
-    // Cancel the timeout which will prevent the delete from occurring
-    if (timerRef.current) clearTimeout(timerRef.current);
-    // Remove the filter that hides the rows that are being deleted
-    setFilters((prevFilters) =>
-      prevFilters.filter(({ type }) => type !== '_exclusion_projection'),
-    );
-  }
-
-  const resetRowSelection = useResetTableRowSelection();
-
   const [deleteCompanies] = useDeleteManyCompaniesMutation({
     refetchQueries: [
       getOperationName(GET_COMPANIES) ?? '',
@@ -43,45 +20,10 @@ export function TableActionBarButtonDeleteCompanies({
     ],
   });
 
-  const [_, setFilters] = useRecoilScopedState(
-    filtersScopedState,
-    TableContext,
-  );
-
-  async function handleDeleteClick() {
-    // Show a snackbar with a countdown
-    enqueueSnackBar(`${selectedRowIds.length} deleted elements`, {
-      icon: <IconTrash />,
-      duration,
-      onClose: handleCloseSnackbar,
-      cancelText: 'Cancel',
-    });
-
-    // Delete the rows after the countdown
-    timerRef.current = setTimeout(() => {
-      const rowIdsToDelete = selectedRowIds;
-
-      resetRowSelection();
-
-      deleteCompanies({
-        variables: {
-          ids: rowIdsToDelete,
-        },
-      });
-      // ensure we delete the rows before the snackbar disappears
-    }, duration - 100);
-
-    // Hide the rows that are being deleted
-    resetRowSelection();
-    const deletedElementsFilter: Filter[] = selectedRowIds.map((id) => ({
-      field: 'id',
-      type: '_exclusion_projection',
-      value: id,
-      displayValue: id,
-      operand: 'is-not',
-    }));
-    setFilters((prevFilters) => [...prevFilters, ...deletedElementsFilter]);
-  }
+  const { handleDeleteClick } = useDeleteItems({
+    timerRef,
+    handleDeleteItems: deleteCompanies,
+  });
 
   return (
     <EntityTableActionBarButton
